pyproject.toml
[tool.poetry]
name = "zoish"
version = "1.63.0"
description = "This project uses shapely values for selecting Top n features compatible with scikit learn pipeline"
authors = ["drhosseinjavedani <h.javedani@gmail.com>"]
homepage = "https://github.com/drhosseinjavedani/zoish"
license = "BSD 2-Clause License"

readme = "README.md"
keywords = [
    "Auto ML",
    "Feature Selection",
    "Pipeline",
    "Machine learning",
    "shap"
]


[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = "^1.4.3"
numpy = "<1.63.0"
scipy = "^1.8.1"
scikit-learn = "^1.1.1"
click = "^8.1.3"
numba = "^0.55.2"
fasttreeshap = "^0.1.2"
shap = "^0.41.0"
xgboost = "^1.6.1"
matplotlib = "^3.5.2"
pip-licenses = "^3.5.4"
feature-engine = "^1.4.1"
optuna = "^2.10.1"
catboost = "^1.0.6"
imblearn = "^0.0"
lightgbm = "^3.3.2"
category-encoders = "^2.5.0"
xgbse = "^0.2.3"
pycox = "^0.2.3"
lohrasb = "^2.1.0"
python-dotenv = "^0.21.0"


[tool.poetry.dev-dependencies]
pytest = ">=6.2.4"
black = "^22.3.0"
flake8 = "^4.0.1"
bump2version = "^1.0.1"
nox = "^2022.1.7"
bumpver = "^2022.1116"
ipykernel = "^6.15.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

